Refactored unit tests structures.
