+++ .vscode/settings.json
@@ -1,6 +1,8 @@
 {
     "rust-analyzer.linkedProjects": [
         ".\\Cargo.toml",
+        ".\\Cargo.toml",
+        ".\\Cargo.toml",
         ".\\Cargo.toml"
     ]
 }
