+++ src/assembler/linker/mod.rs
@@ -42,8 +42,13 @@ use std::collections::HashMap;
 use std::convert::identity;
 use crate::core::instructions::BabyInstruction;
 use super::parser::{LineType, Value, Instruction};
-use super::linker_errors::{LinkingError, TagError};
+use linker_errors::{LinkingError, TagError};
 
+
+/// Contains types for handling errors found during linking. 
+pub mod linker_errors;
+#[cfg(test)]
+mod tests;
 
 /// Links the parsed lines into the corresponding machine code. 
 /// 
@@ -272,218 +277,4 @@ impl UnlinkedData {
             None => Err(tag.to_owned())
         }
     }
-}
-
-#[cfg(test)]
-mod tests {
-
-    use super::*;
-
-    #[test]
-    fn test_inline_tags() {
-        let vec = vec![
-            LineType::Tag("foo".to_owned()),
-            LineType::Instruction(Instruction::Negate(Value::Tag("foo".to_owned()))),
-            LineType::Instruction(Instruction::Negate(Value::Tag("foo".to_owned()))),
-        ];
-        let res = inline_tags(vec);
-        assert_eq!(res.len(), 2);
-        match &res[0] {
-            (Some(tag), UnlinkedData::Instruction(Instruction::Negate(Value::Tag(tag_ref)))) => {
-                assert_eq!(tag, "foo");
-                assert_eq!(tag_ref, "foo")
-            },
-            _ => assert!(false)
-        }
-        match &res[1] {
-            (None, UnlinkedData::Instruction(Instruction::Negate(Value::Tag(tag_ref)))) => 
-                assert_eq!(tag_ref, "foo"),
-            _ => assert!(false)
-        }
-    }
-
-    #[test]
-    fn test_position_tags() {
-        let vec: Vec<(Option<String>, UnlinkedData)> = vec![
-            (Some(format!("foo1")), UnlinkedData::Instruction(Instruction::Negate(Value::Value(5)))),
-            (None, UnlinkedData::Instruction(Instruction::Negate(Value::Value(5)))),
-            (Some(format!("foo2")), UnlinkedData::Instruction(Instruction::Negate(Value::Value(5)))),
-        ];
-
-        let tags = position_tags(&vec);
-        assert_eq!(tags.iter().len(), 2);
-        assert_eq!(tags.get("foo1"), Some(&0));
-        assert_eq!(tags.get("foo2"), Some(&2));
-    }
-
-    #[test]
-    fn test_link_tags_correct() {
-        let tags: HashMap<String, i32> = HashMap::from([("foo".to_owned(), 5)]);
-        let lines: Vec<UnlinkedData> = vec![
-            UnlinkedData::Instruction(Instruction::Jump(Value::Tag("foo".to_owned()))),
-            UnlinkedData::Instruction(Instruction::Jump(Value::Value(5))),
-        ];
-        match link_tags(lines, tags) {
-            Ok(res) => {
-                assert_eq!(res.len(), 2);
-                assert_eq!(res[0], BabyInstruction::Jump(5));
-                assert_eq!(res[1], BabyInstruction::Jump(5));
-            },
-            Err(_) => assert!(false),
-        }
-    }
-
-    #[test]
-    fn test_link_tags_fail() {
-        let tags: HashMap<String, i32> = HashMap::from([("foo".to_owned(), 5)]);
-        let lines: Vec<UnlinkedData> = vec![
-            UnlinkedData::Instruction(Instruction::Jump(Value::Tag("bar".to_owned()))),
-            UnlinkedData::Instruction(Instruction::Jump(Value::Value(5))),
-        ];
-        match link_tags(lines, tags) {
-            Err(e) => {
-                assert_eq!(e, LinkingError::TagError(TagError::UnknownTagName("bar".to_owned())))
-            },
-            Ok(_) => assert!(false),
-        }
-    }
-
-    #[test]
-    fn test_link_parsed_lines_correct() {
-        let lines: Vec<LineType> = vec![
-            LineType::Tag("start".to_owned()),
-            LineType::Instruction(Instruction::Negate(Value::Value(2))),
-            LineType::Instruction(Instruction::Jump(Value::Tag("start".to_owned()))),
-            LineType::Absolute(Value::Value(5)),
-        ];
-        match link_parsed_lines(lines) {
-            Ok(v) => {
-                assert_eq!(v.len(), 3);
-                assert_eq!(v[0], BabyInstruction::Negate(2));
-                assert_eq!(v[1], BabyInstruction::Jump(0));
-                assert_eq!(v[2], BabyInstruction::AbsoluteValue(5));
-            },
-            Err(_) => assert!(false),
-        }
-    }
-
-    #[test]
-    fn test_link_parsed_lines_fail() {
-        let lines: Vec<LineType> = vec![
-            LineType::Tag("start".to_owned()),
-            LineType::Instruction(Instruction::Negate(Value::Value(2))),
-            LineType::Instruction(Instruction::Jump(Value::Tag("foo".to_owned()))),
-            LineType::Absolute(Value::Value(5)),
-        ];
-        match link_parsed_lines(lines) {
-            Err(e) => {
-                assert_eq!(e, LinkingError::TagError(TagError::UnknownTagName("foo".to_owned())))
-            },
-            Ok(_) => assert!(false),
-        }
-    }
-
-}
-
-#[cfg(test)]
-pub mod unlinked_data_tests {
-    
-    use super::*;
-
-    #[test]
-    fn test_get_tag() {
-        let tags: HashMap<String, i32> = HashMap::from([("foo".to_owned(), 5)]);
-        match UnlinkedData::get_tag("foo", &tags) {
-            Ok(v) => assert_eq!(v, 5),
-            Err(_) => assert!(false)
-        }
-        match UnlinkedData::get_tag("bar", &tags) {
-            Err(e) => assert_eq!(e, "bar".to_owned()),
-            Ok(_) => assert!(false)
-        }
-    }
-
-    #[test]
-    fn test_resolve_value() {
-        let tags: HashMap<String, i32> = HashMap::from([("foo".to_owned(), 5)]);
-        match UnlinkedData::resolve_value(&Value::Value(5), &tags) {
-            Ok(v) => assert_eq!(v, 5),
-            Err(_) => assert!(false)
-        }
-        match UnlinkedData::resolve_value(&Value::Tag("foo".to_owned()), &tags) {
-            Ok(v) => assert_eq!(v, 5),
-            Err(_) => assert!(false)
-        }
-        match UnlinkedData::resolve_value(&Value::Tag("bar".to_owned()), &tags) {
-            Err(e) => assert_eq!(e, "bar".to_owned()),
-            Ok(_) => assert!(false)
-        }
-    }
-
-    #[test]
-    fn test_resolve_absolute_value() {
-        let tags: HashMap<String, i32> = HashMap::from([("foo".to_owned(), 5)]);
-        match UnlinkedData::resolve_absolute_value(&Value::Value(5), &tags) {
-            Ok(v) => assert_eq!(v, BabyInstruction::AbsoluteValue(5)),
-            Err(_) => assert!(false)
-        }
-        match UnlinkedData::resolve_absolute_value(&Value::Tag("foo".to_owned()), &tags) {
-            Ok(v) => assert_eq!(v, BabyInstruction::AbsoluteValue(5)),
-            Err(_) => assert!(false)
-        }
-        match UnlinkedData::resolve_absolute_value(&Value::Tag("bar".to_owned()), &tags) {
-            Err(e) => assert_eq!(e, TagError::UnknownTagName("bar".to_owned())),
-            Ok(_) => assert!(false)
-        }
-    }
-
-    fn get_litteral_value_instruction(value: i32, result: u16) -> Vec<(Instruction, Result<BabyInstruction, TagError>)> {
-        vec![
-            (Instruction::Jump(Value::Value(value)), Ok(BabyInstruction::Jump(result))),
-            (Instruction::RelativeJump(Value::Value(value)), Ok(BabyInstruction::RelativeJump(result))),
-            (Instruction::Negate(Value::Value(value)), Ok(BabyInstruction::Negate(result))),
-            (Instruction::Store(Value::Value(value)), Ok(BabyInstruction::Store(result))),
-            (Instruction::Subtract(Value::Value(value)), Ok(BabyInstruction::Subtract(result))),
-            (Instruction::Test, Ok(BabyInstruction::SkipNextIfNegative)),
-            (Instruction::Stop, Ok(BabyInstruction::Stop)),
-        ]
-    }
-
-    fn get_tag_ref_instructions(value: &str, res: u16) -> Vec<(Instruction, Result<BabyInstruction, TagError>)> {
-        vec![
-            (Instruction::Jump(Value::Tag(String::from(value))), Ok(BabyInstruction::Jump(res))),
-            (Instruction::RelativeJump(Value::Tag(String::from(value))), Ok(BabyInstruction::RelativeJump(res))),
-            (Instruction::Negate(Value::Tag(String::from(value))), Ok(BabyInstruction::Negate(res))),
-            (Instruction::Store(Value::Tag(String::from(value))), Ok(BabyInstruction::Store(res))),
-            (Instruction::Subtract(Value::Tag(String::from(value))), Ok(BabyInstruction::Subtract(res))),
-        ]
-    }
-
-    fn get_tag_ref_instructions_err(value: &str, err: TagError) -> Vec<(Instruction, Result<BabyInstruction, TagError>)> {
-        vec![
-            (Instruction::Jump(Value::Tag(String::from(value))), Err(err.clone())),
-            (Instruction::RelativeJump(Value::Tag(String::from(value))), Err(err.clone())),
-            (Instruction::Negate(Value::Tag(String::from(value))), Err(err.clone())),
-            (Instruction::Store(Value::Tag(String::from(value))), Err(err.clone())),
-            (Instruction::Subtract(Value::Tag(String::from(value))), Err(err.clone())),
-        ]
-    }
-
-    #[test]
-    fn test_resolve_instruction() {
-        let tags: HashMap<String, i32> = HashMap::from([("foo".to_owned(), 5)]);
-
-        get_litteral_value_instruction(5, 5).iter().for_each(|(i, res)| {
-            assert_eq!(UnlinkedData::resolve_instruction(i, &tags), *res);
-        });
-
-        get_tag_ref_instructions("foo", 5).iter().for_each(|(i, res)| {
-            assert_eq!(UnlinkedData::resolve_instruction(i, &tags), *res);
-        });
-
-        get_tag_ref_instructions_err("bar", TagError::UnknownTagName("bar".to_owned())).iter().for_each(|(i, res)| {
-            assert_eq!(UnlinkedData::resolve_instruction(i, &tags), *res);
-        });
-    }
-
 }
