+++ src/assembler/mod.rs
@@ -178,12 +178,8 @@ use errors::AssemblyError;
 
 /// Contains types and functionality for parsing Baby asm. 
 pub mod parser;
-/// Contains types for handling errors during parsing. 
-pub mod parse_errors;
 /// Contains types and functionality for linking asm post parsing.  
 pub mod linker;
-/// Contains types for handling errors found during linking. 
-pub mod linker_errors;
 /// Contains types for handling errors found during the overall 
 /// assembling process. 
 pub mod errors;
