+++ src/assembler/parser/mod.rs
@@ -34,7 +34,7 @@
 //! ```
 //! 
 
-use super::parse_errors::{
+use parse_errors::{
     LineParseError,
     TagError,
     AbsoluteError,
@@ -42,6 +42,11 @@ use super::parse_errors::{
     ValueParseError,
 };
 
+
+/// Contains types for handling errors during parsing. 
+pub mod parse_errors;
+#[cfg(test)]
+mod tests;
 
 /// Represents the possible nations for a value. 
 #[derive(Clone, Debug, PartialEq)]
@@ -445,244 +450,4 @@ pub fn parse_instruction_ogn(instruction: String) -> Result<LineType, LineParseE
         Ok(v) => Ok(LineType::Instruction(v)),
         Err(e) => Err(LineParseError::InstructionError(e))
     }
-}
-
-#[cfg(test)]
-mod tests {
-
-    use super::*;
-
-    #[test]
-    fn test_strip_comments() {
-        assert_eq!(strip_comments("jmp 0x3; hello world"), "jmp 0x3".to_owned());
-        assert_eq!(strip_comments("jmp 0x3"), "jmp 0x3".to_owned());
-    }
-
-    #[test]
-    fn test_parse_tag() {
-        assert_eq!(parse_tag("  test  ".to_owned()), Ok(LineType::Tag("test".to_owned())));
-        assert_eq!(parse_tag("test foo".to_owned()), Err(LineParseError::TagError(TagError::TagNameWhitespace("test foo".to_owned()))));
-    }
-
-    fn get_absolute() -> Vec<(String, Result<LineType, LineParseError>)> {
-        let v = 10;
-        vec![
-            ("  0xA  ".to_owned(), Ok(LineType::Absolute(Value::Value(v)))),
-            ("  0d10  ".to_owned(), Ok(LineType::Absolute(Value::Value(v)))),
-            ("  0o12  ".to_owned(), Ok(LineType::Absolute(Value::Value(v)))),
-            ("  0b1010  ".to_owned(), Ok(LineType::Absolute(Value::Value(v)))),
-            ("  $tag  ".to_owned(), Ok(LineType::Absolute(Value::Tag("tag".to_owned())))),
-            (
-                "dasdasdas".to_owned(), 
-                Err(LineParseError::AbsoluteError(AbsoluteError::ValueError(ValueParseError::InvalidValue("dasdasdas".to_owned()))))
-            ),
-        ]
-    }
-
-    fn get_absolute_err() -> Vec<(String, Result<LineType, LineParseError>)> {
-        vec![
-            ("0xK".to_owned(), Err(LineParseError::AbsoluteError(AbsoluteError::ValueError(ValueParseError::InvalidHex("K".to_owned()))))),
-            ("0d1J".to_owned(), Err(LineParseError::AbsoluteError(AbsoluteError::ValueError(ValueParseError::InvalidDecimal("1J".to_owned()))))),
-            ("0o9".to_owned(), Err(LineParseError::AbsoluteError(AbsoluteError::ValueError(ValueParseError::InvalidOctal("9".to_owned()))))),
-            ("0b10a10".to_owned(), Err(LineParseError::AbsoluteError(AbsoluteError::ValueError(ValueParseError::InvalidBinary("10a10".to_owned()))))),
-            ("$ta g".to_owned(), Err(LineParseError::AbsoluteError(AbsoluteError::ValueError(ValueParseError::InvalidTagName("ta g".to_owned()))))),
-        ]
-    }
-
-    fn get_instructions() -> Vec<(String, Result<LineType, LineParseError>)> {
-        vec![
-            ("  jmp   0xA  ".to_owned(), Ok(LineType::Instruction(Instruction::Jump(Value::Value(10))))),
-            ("  jrp   0xA  ".to_owned(), Ok(LineType::Instruction(Instruction::RelativeJump(Value::Value(10))))),
-            ("  ldn   0xA  ".to_owned(), Ok(LineType::Instruction(Instruction::Negate(Value::Value(10))))),
-            ("  sto   0xA  ".to_owned(), Ok(LineType::Instruction(Instruction::Store(Value::Value(10))))),
-            ("  sub   0xA  ".to_owned(), Ok(LineType::Instruction(Instruction::Subtract(Value::Value(10))))),
-            ("  cmp  ".to_owned(), Ok(LineType::Instruction(Instruction::Test))),
-            ("  stp  ".to_owned(), Ok(LineType::Instruction(Instruction::Stop))),
-            (
-                "sadasdasd".to_owned(), 
-                Err(LineParseError::InstructionError(InstructionError::UnkownInstruction("sadasdasd".to_owned())))
-            ),
-        ]
-    }
-
-    fn get_ogn_instructions() -> Vec<(String, Result<LineType, LineParseError>)> {
-        vec![
-            ("  0xA  , Cl  ".to_owned(), Ok(LineType::Instruction(Instruction::Jump(Value::Value(10))))),
-            ("  Add   0xA  , Cl  ".to_owned(), Ok(LineType::Instruction(Instruction::RelativeJump(Value::Value(10))))),
-            ("  -  0xA  , C  ".to_owned(), Ok(LineType::Instruction(Instruction::Negate(Value::Value(10))))),
-            ("  c, 0xA  ".to_owned(), Ok(LineType::Instruction(Instruction::Store(Value::Value(10))))),
-            ("  SUB   0xA  ".to_owned(), Ok(LineType::Instruction(Instruction::Subtract(Value::Value(10))))),
-            ("  Test  ".to_owned(), Ok(LineType::Instruction(Instruction::Test))),
-            ("  Stop  ".to_owned(), Ok(LineType::Instruction(Instruction::Stop))),
-            (
-                "sadasdasd".to_owned(), 
-                Err(LineParseError::InstructionError(InstructionError::UnkownInstruction("sadasdasd".to_owned())))
-            ),
-        ]
-    }
-
-    #[test]
-    fn test_parse_absolute() {
-        get_absolute().iter().for_each(|(v, r)| {
-            assert_eq!(parse_absolute(v.to_string()), *r);
-        });
-    }
-
-    #[test]
-    fn test_parse_absolute_err() {
-        get_absolute_err().iter().for_each(|(v, r)| {
-            assert_eq!(parse_absolute(v.to_string()), *r);
-        });
-    }
-
-    #[test]
-    fn test_parse_instruction() {
-        get_instructions().iter().for_each(|(v, r)| {
-            assert_eq!(parse_instruction(v.to_string()), *r);
-        });
-    }
-
-    #[test]
-    fn test_parse_instruction_ogn() {
-        get_ogn_instructions().iter().for_each(|(v, r)| {
-            assert_eq!(parse_instruction_ogn(v.to_string()), *r)
-        });
-    }
-}
-
-#[cfg(test)]
-mod instruction_test {
-
-    use super::*;
-
-    fn get_intrsuctions<'a>() -> Vec<(&'a str, Result<Instruction, InstructionError>)> {
-        let v = 10;
-        vec![
-            ("jmp 0xA", Ok(Instruction::Jump(Value::Value(v)))),
-            ("jrp 0xA", Ok(Instruction::RelativeJump(Value::Value(v)))),
-            ("ldn 0xA", Ok(Instruction::Negate(Value::Value(v)))),
-            ("sto 0xA", Ok(Instruction::Store(Value::Value(v)))),
-            ("sub 0xA", Ok(Instruction::Subtract(Value::Value(v)))),
-            ("cmp", Ok(Instruction::Test)),
-            ("stp", Ok(Instruction::Stop)),
-            ("sadasdasd", Err(InstructionError::UnkownInstruction("sadasdasd".to_owned()))),
-        ]
-    }
-
-    fn get_ogn_intrsuctions<'a>() -> Vec<(&'a str, Result<Instruction, InstructionError>)> {
-        let v = 10;
-        vec![
-            ("0xA, Cl", Ok(Instruction::Jump(Value::Value(v)))),
-            ("Add 0xA, Cl", Ok(Instruction::RelativeJump(Value::Value(v)))),
-            ("-0xA, C", Ok(Instruction::Negate(Value::Value(v)))),
-            ("c, 0xA", Ok(Instruction::Store(Value::Value(v)))),
-            ("SUB 0xA", Ok(Instruction::Subtract(Value::Value(v)))),
-            ("Test", Ok(Instruction::Test)),
-            ("Stop", Ok(Instruction::Stop)),
-            ("sadasdasd", Err(InstructionError::UnkownInstruction("sadasdasd".to_owned()))),
-        ]
-    }
-
-    fn get_instructions_with_operands() -> Vec<(Instruction, Value)> {
-        vec![
-            (Instruction::Jump(Value::Value(10)), Value::Value(10)),
-            (Instruction::RelativeJump(Value::Value(10)), Value::Value(10)),
-            (Instruction::Negate(Value::Value(10)), Value::Value(10)),
-            (Instruction::Store(Value::Value(10)), Value::Value(10)),
-            (Instruction::Subtract(Value::Value(10)), Value::Value(10)),
-            (Instruction::Test, Value::Value(0)),
-            (Instruction::Stop, Value::Value(0)),
-        ]
-    }
-
-    fn get_instructions_with_description() -> Vec<(Instruction, String)> {
-        vec![
-            (Instruction::Jump(Value::Value(10)), "jump".to_owned()),
-            (Instruction::RelativeJump(Value::Value(10)), "relative jump".to_owned()),
-            (Instruction::Negate(Value::Value(10)), "negate".to_owned()),
-            (Instruction::Store(Value::Value(10)), "store".to_owned()),
-            (Instruction::Subtract(Value::Value(10)), "subtract".to_owned()),
-            (Instruction::Test, "test".to_owned()),
-            (Instruction::Stop, "stop".to_owned()),
-        ]
-    }
-
-    #[test]
-    fn test_instruction_parse() {
-        get_intrsuctions().iter().for_each(|(i, r)| {
-            assert_eq!(Instruction::parse(i), *r);
-        });
-    }
-
-    #[test]
-    fn test_instruction_parse_ogn() {
-        get_ogn_intrsuctions().iter().for_each(|(i, r)| {
-            assert_eq!(Instruction::parse_ogn(i), *r);
-        });
-    }
-
-    #[test]
-    fn test_get_operand() {
-        get_instructions_with_operands().iter().for_each(|(i, r)| {
-            assert_eq!(i.get_operand(), *r);
-        });
-    }
-
-    #[test]
-    fn test_describe() {
-        get_instructions_with_description().iter().for_each(|(i, r)| {
-            assert_eq!(i.describe(), *r);
-        });
-    }
-}
-
-#[cfg(test)]
-mod value_tests {
-
-    use super::*;
-
-    #[test]
-    fn test_parse_hex() {
-        assert_eq!(Value::parse_hex("A".to_owned()), Ok(Value::Value(10)));
-        assert_eq!(Value::parse_hex("-A".to_owned()), Ok(Value::Value(-10)));
-        assert_eq!(Value::parse_hex("K".to_owned()), Err(ValueParseError::InvalidHex("K".to_owned())));
-    }
-
-    #[test]
-    fn test_parse_decimal() {
-        assert_eq!(Value::parse_decimal("10".to_owned()), Ok(Value::Value(10)));
-        assert_eq!(Value::parse_decimal("-10".to_owned()), Ok(Value::Value(-10)));
-        assert_eq!(Value::parse_decimal("K".to_owned()), Err(ValueParseError::InvalidDecimal("K".to_owned())));
-    }
-
-    #[test]
-    fn test_parse_octal() {
-        assert_eq!(Value::parse_octal("12".to_owned()), Ok(Value::Value(10)));
-        assert_eq!(Value::parse_octal("-12".to_owned()), Ok(Value::Value(-10)));
-        assert_eq!(Value::parse_octal("K".to_owned()), Err(ValueParseError::InvalidOctal("K".to_owned())));
-    }
-
-    #[test]
-    fn test_parse_binary() {
-        assert_eq!(Value::parse_binary("1010".to_owned()), Ok(Value::Value(10)));
-        assert_eq!(Value::parse_binary("-1010".to_owned()), Ok(Value::Value(-10)));
-        assert_eq!(Value::parse_binary("K".to_owned()), Err(ValueParseError::InvalidBinary("K".to_owned())));
-    }
-
-    #[test]
-    fn test_parse_tag() {
-        assert_eq!(Value::parse_tag_name("tag".to_owned()), Ok(Value::Tag("tag".to_owned())));
-        assert_eq!(Value::parse_tag_name("ta g".to_owned()), Err(ValueParseError::InvalidTagName("ta g".to_owned())))
-    }
-
-    #[test]
-    fn test_parse() {
-        assert_eq!(Value::parse("  0xA  "), Ok(Value::Value(10)));
-        assert_eq!(Value::parse("  0d10  "), Ok(Value::Value(10)));
-        assert_eq!(Value::parse("  0o12  "), Ok(Value::Value(10)));
-        assert_eq!(Value::parse("  0b1010  "), Ok(Value::Value(10)));
-        assert_eq!(Value::parse("  $foo  "), Ok(Value::Tag("foo".to_owned())));
-        assert_eq!(Value::parse("  sadfdsfsda  "), Err(ValueParseError::InvalidValue("sadfdsfsda".to_owned())));
-    }
-
 }
